+++ packages/path_provider/path_provider_foundation/darwin/Classes/messages.g.swift
@@ -44,6 +44,7 @@ enum DirectoryType: Int {
   case downloads = 2
   case library = 3
   case temp = 4
+  case applicationCache = 5
 }
 /// Generated protocol from Pigeon that represents a handler of messages from Flutter.
 protocol PathProviderApi {