+++ packages/path_provider/path_provider_linux/lib/src/path_provider_linux.dart
@@ -72,6 +72,18 @@ class PathProviderLinux extends PathProviderPlatform {
   }
 
   @override
+  Future<String?> getApplicationCachePath() async {
+    final Directory directory =
+        Directory(path.join(xdg.cacheHome.path, await _getId()));
+    if (directory.existsSync()) {
+      return directory.path;
+    }
+
+    await directory.create(recursive: true);
+    return directory.path;
+  }
+
+  @override
   Future<String?> getDownloadsPath() {
     return Future<String?>.value(xdg.getUserDirectory('DOWNLOAD')?.path);
   }