+++ packages/path_provider/path_provider_platform_interface/lib/src/enums.dart
@@ -50,4 +50,7 @@ enum StorageDirectory {
 
   /// Holds user-created documents. See https://developer.android.com/reference/android/os/Environment.html#DIRECTORY_DOCUMENTS.
   documents,
+
+  /// Holds application cache files.
+  cache,
 }