+++ packages/path_provider/path_provider_foundation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.3.0
+
+* Adds getApplicationCachePath() for storing app-specific cache files.
+
 ## 2.2.4
 
 * Updates to the latest version of `pigeon`.